requirements.txt
absl-py==2.3.1
accelerate==1.10.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
astor==0.8.1
asttokens==3.0.1
attrs==25.4.0
binpacking==1.5.2
blake3==1.0.8
cachetools==6.2.1
causal_conv1d==1.5.3.post1
cbor2==5.7.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.1
comm==0.2.3
compressed-tensors==0.11.0
conda-pack==0.8.1
contourpy==1.3.3
cupy-cuda12x==13.6.0
cycler==0.12.1
datasets==3.3.0
debugpy==1.8.17
decorator==5.2.1
deepspeed==0.18.1
depyf==0.19.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fastrlock==0.8.3
filelock==3.20.0
fla-core==0.3.2
flash-linear-attention==0.3.2
flash_attn @ file:///mnt/cxzx/share/envs/flash_attn-2.8.3%2Bcu12torch2.8cxx11abiTRUE-cp312-cp312-linux_x86_64.whl#sha256=f25da18657a87fc83dc1bfb8b7751b82246e9db355510226b674fd437c34b5fb
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2024.9.0
future==1.0.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
h11==0.16.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
iniconfig==2.3.0
interegular==0.3.3
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lark==1.2.2
liger_kernel==0.6.3
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
lmdb==1.7.5
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistral_common==1.8.5
mmengine==0.10.7
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==2.5.0
openai-harmony==0.0.4
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
outlines_core==0.2.11
packaging==25.0
pandas==2.3.3
parso==0.8.5
partial-json-parser==0.2.1.1.post6
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-extra-types==2.10.6
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.50.1
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
ring-flash-attn==0.1.8
rpds-py==0.27.1
safetensors==0.6.2
scipy==1.16.2
sentencepiece==0.2.1
sentry-sdk==2.42.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
torchcodec==0.8.1
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.4.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.11.0
wandb==0.22.3
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
xformers==0.0.32.post1
xgrammar==0.1.25
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
